Exception check

+++ entrypoint.sh
@@ -165,18 +165,19 @@ write_step_summary() {
       gh_warn "Log file for '${MCIX_CMD_NAME}' not found" "Continuing without failing the action."
   fi
 
-  for file in $MCIX_LOG_DIR/exception.*.log; do
-    echo " - $file"
-    {
-      # Display the contents of the mcix command's log file. (collapsed by default)
-      echo '<details>'
-      echo "<summary>Exception Log - $file</summary>"
-      echo # A blank line after the <summary> tag is required by GitHub to format the content correctly
-      echo '```'
-      cat $file
-      echo '```'
-      echo '</details>'
-    } >>"$GITHUB_STEP_SUMMARY"
+  for file in "$MCIX_LOG_DIR/exception.*.log"; do
+    if [ -f "$file" ]; then
+      {
+        # Display the contents of the mcix command's log file. (collapsed by default)
+        echo '<details>'
+        echo "<summary>Exception Log - $file</summary>"
+        echo # A blank line after the <summary> tag is required by GitHub to format the content correctly
+        echo '```'
+        cat $file
+        echo '```'
+        echo '</details>'
+      } >>"$GITHUB_STEP_SUMMARY"
+    fi
   done
 }
